+++ ShopmeWebParent/ShopmeBackEnd/src/main/java/com/shopme/admin/security/ShopmeUserDetails.java
@@ -25,7 +25,7 @@ public class ShopmeUserDetails implements UserDetails {
         List<SimpleGrantedAuthority> authorities = new ArrayList<>();
 
         for (Role role : roles) {
-            authorities.add(new SimpleGrantedAuthority(role.getDescription()));
+            authorities.add(new SimpleGrantedAuthority(role.getName()));
         }
 
         return authorities;
