+++ ShopmeWebParent/ShopmeBackEnd/src/main/java/com/shopme/admin/security/WebSecurityConfig.java
@@ -34,7 +34,9 @@ public class WebSecurityConfig {
 
     @Bean
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        http.authorizeHttpRequests()
+        http.authorizeHttpRequests().requestMatchers("/users/**")
+                .hasAuthority("Admin").and()
+                .authorizeHttpRequests()
                 .requestMatchers("/images/**", "/js/**", "/webjars/**")
                 .permitAll().and().authorizeHttpRequests().requestMatchers("/**")
                 .authenticated().and()
