--- ShopmeWebParent/ShopmeBackEnd/src/main/java/com/shopme/admin/user/UserController.java
@@ -1,158 +1,0 @@
-package com.shopme.admin.user;
-
-import com.shopme.admin.FileUploadUtil;
-import com.shopme.admin.user.export.UserCsvExporter;
-import com.shopme.admin.user.export.UserExcelExporter;
-import com.shopme.admin.user.export.UserPdfExporter;
-import com.shopme.common.entity.Role;
-import com.shopme.common.entity.User;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.data.domain.Page;
-import org.springframework.data.repository.query.Param;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.util.StringUtils;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.servlet.mvc.support.RedirectAttributes;
-
-import java.io.IOException;
-import java.util.List;
-import java.util.Objects;
-
-@Controller
-public class UserController {
-
-    private final UserService userService;
-
-    public UserController(UserService userService) {
-        this.userService = userService;
-    }
-
-    @GetMapping("/users")
-    public String listFirstPage(Model model) {
-        return listByPage(1, model, "firstName", "asc", null);
-    }
-
-    @GetMapping("/users/page/{pageNum}")
-    public String listByPage(@PathVariable(name = "pageNum") int pageNum, Model model,
-                             @Param("sortField") String sortField,
-                             @Param("sortDir") String sortDir,
-                             @Param("keyword") String keyword) {
-        Page<User> page = userService.listByPage(pageNum, sortField, sortDir, keyword);
-        List<User> userList = page.getContent();
-        long startCount = (long) (pageNum - 1) * UserService.USERS_PER_PAGE + 1;
-        long endCount = startCount + UserService.USERS_PER_PAGE - 1;
-        if (endCount > page.getTotalElements()) {
-            endCount = page.getTotalElements();
-        }
-        String reverseSortDir = sortDir.equals("asc") ? "desc" : "asc";
-        model.addAttribute("currentPage", pageNum);
-        model.addAttribute("totalPages", page.getTotalPages());
-        model.addAttribute("startCount", startCount);
-        model.addAttribute("endCount", endCount);
-        model.addAttribute("totalItems", page.getTotalElements());
-        model.addAttribute("listUsers", userList);
-        model.addAttribute("sortField", sortField);
-        model.addAttribute("sortDir", sortDir);
-        model.addAttribute("reverseSortDir", reverseSortDir);
-        model.addAttribute("keyword", keyword);
-        return "users";
-    }
-
-    @GetMapping("/users/new")
-    public String createNewUser(Model model) {
-        List<Role> roleList = userService.roleList();
-        User user = new User();
-        user.setEnabled(true);
-        model.addAttribute("pageTitle", "Create New User");
-        model.addAttribute("user", user);
-        model.addAttribute("listRoles", roleList);
-        return "user_form";
-    }
-
-    @PostMapping("/users/save")
-    public String saveUser(User user, RedirectAttributes redirectAttributes,
-                           @RequestParam("image")MultipartFile multipartFile) throws IOException {
-        if (!multipartFile.isEmpty()) {
-            String fileName = StringUtils.cleanPath(Objects.requireNonNull(multipartFile.getOriginalFilename()));
-            user.setPhotos(fileName);
-            User savedUser = userService.save(user);
-            String uploadDir = "user-photos/" + savedUser.getId();
-            FileUploadUtil.cleanDir(uploadDir);
-            FileUploadUtil.saveFile(uploadDir, fileName, multipartFile);
-         } else {
-            if (user.getPhotos().isEmpty()) user.setPhotos(null);
-            userService.save(user);
-        }
-        redirectAttributes.addFlashAttribute("message", "The user has been saved successfully");
-        return getRedirectURLtoAffectedUser(user);
-    }
-
-    private static String getRedirectURLtoAffectedUser(User user) {
-        String firstPartOfEmail = user.getEmail().split("@")[0];
-        return "redirect:/users/page/1?sortField=id&sortDir=asc&keyword=" + firstPartOfEmail;
-    }
-
-    @GetMapping("/users/edit/{id}")
-    public String editUser(@PathVariable(name = "id") Integer id,
-                           Model model, RedirectAttributes redirectAttributes) {
-        try {
-            User user = userService.getById(id);
-            List<Role> roleList = userService.roleList();
-            model.addAttribute("user", user);
-            model.addAttribute("listRoles", roleList);
-            model.addAttribute("pageTitle", "Edit User ID: " + id);
-            return "user_form";
-        } catch (UserNotFoundException e) {
-            redirectAttributes.addFlashAttribute("message", e.getMessage());
-            return "redirect:/users";
-        }
-    }
-
-    @GetMapping("/users/delete/{id}")
-    public String deleteUser(@PathVariable("id") Integer id, RedirectAttributes redirectAttributes) {
-        try {
-            userService.deleteById(id);
-            redirectAttributes
-                    .addFlashAttribute("message", "The user has been deleted successfully");
-        } catch (UserNotFoundException e) {
-            redirectAttributes.addFlashAttribute("message", e.getMessage());
-        }
-        return "redirect:/users";
-    }
-
-    @GetMapping("/users/{id}/enabled/{status}")
-    public String enableUser(@PathVariable("id") Integer id, @PathVariable("status") boolean enabled,
-                             RedirectAttributes redirectAttributes) {
-        userService.updateUserEnabledStatus(id, enabled);
-        String status = enabled ? "enabled" : "disabled";
-        String message = String.format("The user ID %s has been %s", id, status);
-        redirectAttributes.addFlashAttribute("message", message);
-        return "redirect:/users";
-    }
-
-    @GetMapping("/users/export/csv")
-    public void exportUsersToCsv(HttpServletResponse response) throws IOException {
-        List<User> userList = userService.listAll();
-        UserCsvExporter csvExporter = new UserCsvExporter();
-        csvExporter.export(userList, response);
-    }
-
-    @GetMapping("/users/export/excel")
-    public void exportUsersToExcel(HttpServletResponse response) throws IOException {
-        List<User> userList = userService.listAll();
-        UserExcelExporter excelExporter = new UserExcelExporter();
-        excelExporter.export(userList, response);
-    }
-
-    @GetMapping("/users/export/pdf")
-    public void exportUsersToPdf(HttpServletResponse response) throws IOException {
-        List<User> userList = userService.listAll();
-        UserPdfExporter pdfExporter = new UserPdfExporter();
-        pdfExporter.export(userList, response);
-    }
-}
