+++ ShopmeWebParent/ShopmeBackEnd/src/main/java/com/shopme/admin/user/controller/UserRestController.java
@@ -1,0 +1,23 @@
+package com.shopme.admin.user.controller;
+
+import com.shopme.admin.user.UserService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.repository.query.Param;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class UserRestController {
+    private final UserService userService;
+
+    @Autowired
+    public UserRestController(UserService userService) {
+        this.userService = userService;
+    }
+
+    @PostMapping("/users/check_email")
+    public String checkDuplicateEmail(@Param("id") Integer id, @Param("email") String email) {
+        return userService.isEmailUnique(id, email) ? "OK" : "Duplicated";
+    }
+
+}
