+++ ShopmeWebParent/ShopmeBackEnd/src/main/resources/templates/index.html
@@ -1,12 +1,18 @@
 <!DOCTYPE html>
 <html xmlns:th="https://www.thymeleaf.org"
->
+      xmlns:sec="http://www.thymeleaf.org/extras/spring-security" lang="en">
 <head th:replace="~{fragments :: page_head('Home page - Shopeme Admin')}"/>
 <body>
 <div class="container-fluid">
     <div th:replace="~{navigation :: menu}"></div>
-    <div th:replace="~{fragments :: footer}">
+    <div>
+        <h2>Shopme Control Panel</h2>
+        Welcome <a th:href="@{/account}"
+                   sec:authentication="principal.fullName"></a>
+        &nbsp;
+        <span sec:authentication="principal.authorities"></span>
     </div>
+    <div th:replace="~{fragments :: footer}"></div>
     <script type="text/javascript">
 
     </script>
