+++ ShopmeWebParent/ShopmeBackEnd/src/main/resources/templates/navigation.html
@@ -15,34 +15,55 @@
         </button>
         <div class="collapse navbar-collapse" id="topNavbar">
             <ul class="navbar-nav">
-                <li class="nav-item">
+                <li class="nav-item" sec:authorize="hasAuthority('Admin')">
                     <a class="nav-link" th:href="@{/users}">Users</a>
                 </li>
+                <th:block sec:authorize="hasAnyAuthority('Admin', 'Editor')">
                 <li class="nav-item">
                     <a class="nav-link" th:href="@{/categories}">Categories</a>
                 </li>
+                </th:block>
+                <th:block sec:authorize="hasAnyAuthority('Admin', 'Editor')">
                 <li class="nav-item">
                     <a class="nav-link" th:href="@{/brands}">Brands</a>
                 </li>
+                </th:block>
+                <th:block sec:authorize="hasAnyAuthority('Admin', 'Salesperson', 'Editor', 'Shipper')">
                 <li class="nav-item">
                     <a class="nav-link" th:href="@{/products}">Products</a>
                 </li>
+                </th:block>
+                <th:block sec:authorize="hasAnyAuthority('Admin', 'Salesperson')">
                 <li class="nav-item">
                     <a class="nav-link" th:href="@{/customers}">Customers</a>
                 </li>
+                </th:block>
+                <th:block sec:authorize="hasAnyAuthority('Admin', 'Salesperson')">
                 <li class="nav-item">
                     <a class="nav-link" th:href="@{/shipping}">Shipping</a>
                 </li>
+                </th:block>
+                <th:block sec:authorize="hasAnyAuthority('Admin', 'Salesperson', 'Shipper')">
+                    <li class="nav-item">
+                        <a class="nav-link" th:href="@{/orders}">Orders</a>
+                    </li>
+                </th:block>
+                <th:block sec:authorize="hasAnyAuthority('Admin', 'Salesperson')">
                 <li class="nav-item">
                     <a class="nav-link" th:href="@{/report}">Sales Report</a>
                 </li>
+                </th:block>
+                <th:block sec:authorize="hasAnyAuthority('Admin', 'Editor')">
                 <li class="nav-item">
                     <a class="nav-link" th:href="@{/articles}">Articles</a>
                 </li>
+                </th:block>
+                <th:block sec:authorize="hasAnyAuthority('Admin', 'Editor')">
                 <li class="nav-item">
                     <a class="nav-link" th:href="@{/menus}">Menus</a>
                 </li>
-                <li class="nav-item">
+                </th:block>
+                <li class="nav-item"  sec:authorize="hasAuthority('Admin')">
                     <a class="nav-link" th:href="@{/settings}">Settings</a>
                 </li>
                 <li>
