--- ShopmeWebParent/ShopmeBackEnd/src/main/resources/templates/user_form.html
@@ -1,160 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head th:replace="~{fragments :: page_head(${pageTitle})}">
-<body>
-<div class="container-fluid">
-    <div th:replace="~{navigation :: menu}"></div>
-    <div>
-        <h2>Manage Users | [[${pageTitle}]]</h2>
-    </div>
-    <form th:action="@{/users/save}" method="post" style="max-width: 700px; margin: 0 auto"
-          enctype="multipart/form-data"
-          th:object="${user}" onsubmit="return checkEmailUnique(this);">
-        <input type="hidden" th:field="*{id}">
-        <div class="border border-secondary rounded p-3">
-            <div class="form-group row">
-                <label class="col-sm-4 col-form-label">E-mail:</label>
-                <div class="col-sm-8">
-                    <input type="email" class="form-control" th:field="*{email}" required minlength="7" maxlength="128">
-                </div>
-            </div>
-            <div class="form-group row">
-                <label class="col-sm-4 col-form-label">First Name:</label>
-                <div class="col-sm-8">
-                    <input type="text" class="form-control" th:field="*{firstName}" required minlength="2"
-                           maxlength="45">
-                </div>
-            </div>
-            <div class="form-group row">
-                <label class="col-sm-4 col-form-label">Last Name:</label>
-                <div class="col-sm-8">
-                    <input type="text" class="form-control" th:field="*{lastName}" required minlength="2"
-                           maxlength="45">
-                </div>
-            </div>
-            <div class="form-group row">
-                <label class="col-sm-4 col-form-label">Password:</label>
-                <div class="col-sm-8">
-                    <input th:if="${user.id == null}" type="password" class="form-control" th:field="*{password}"
-                           required minlength="8"
-                           maxlength="20">
-                    <input th:if="${user.id != null}" type="password" class="form-control" th:field="*{password}"
-                           minlength="8"
-                           maxlength="20" placeholder="If you don't want to change password leave as is">
-                </div>
-            </div>
-            <div class="form-group row">
-                <label class="col-sm-4 col-form-label">Roles:</label>
-                <div class="col-sm-8">
-                    <th:block th:each="role : ${listRoles}">
-                        <input type="checkbox" th:field="*{roles}"
-                               th:text="${role.name}"
-                               th:value="${role.id}"
-                               class="m-2"
-                        >
-                        - <small>[[${role.description}]]</small>
-                        <br/>
-                    </th:block>
-                </div>
-                <div class="form-group row">
-                    <label class="col-sm-4 col-form-label">Enabled:</label>
-                    <div class="col-sm-8">
-                        <input type="checkbox" th:field="*{enabled}">
-                    </div>
-                </div>
-                <div class="form-group row">
-                    <label class="col-sm-4 col-form-label">Photo:</label>
-                    <div class="col-sm-8">
-                        <input type="hidden" th:field="*{photos}"/>
-                        <input type="file" id="file_image" name="image" accept="image/jpeg, image/png" class="mb-2">
-                        <img id="thumbnail" alt="Photos preview" th:src="@{${user.photosImagePath}}"
-                             class="img-fluid" style="width: 100px"/>
-                    </div>
-                </div>
-            </div>
-            <div class="text-center">
-                <input type="submit" value="Save" class="btn btn-primary m-3"/>
-                <input type="button" value="Cancel" class="btn btn-secondary m-3" id="btn_cancel">
-            </div>
-        </div>
-    </form>
-    <div class="modal fade text-center" id="modal_dialog">
-        <div class="modal-dialog">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h4 class="modal-title" id="modal_title">Warning</h4>
-                    <button type="button" class="btn-close" data-bs-dismiss="modal">&times;</button>
-                </div>
-                <div class="modal-body">
-                    <span id="modal_body"></span>
-                </div>
-
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-danger" data-bs-dismiss="modal">Close</button>
-                </div>
-            </div>
-        </div>
-    </div>
-    <div>
-
-    </div>
-    <div th:replace="~{fragments :: footer}"></div>
-</div>
-</body>
-<script type="text/javascript">
-    $(document).ready(function () {
-        $("#btn_cancel").on("click", function () {
-            window.location = "[[@{/users}]]";
-        });
-
-        $("#file_image").change(function () {
-            file_size = this.files[0].size
-            alert("File size: " + file_size);
-            if (file_size > 1048576) {
-                this.setCustomValidity("You must choose an image less than 1MB!");
-                this.reportValidity()
-            } else {
-                this.setCustomValidity("");
-                showImageThumbnail(this);
-            }
-        })
-    });
-
-    function showImageThumbnail(fileInput) {
-        var file = fileInput.files[0];
-        var reader = new FileReader();
-        reader.onload = function (e) {
-            $("#thumbnail").attr("src", e.target.result)
-        };
-
-        reader.readAsDataURL(file);
-    }
-
-    function checkEmailUnique(form) {
-        url = "[[@{/users/check_email}]]";
-        userId = $("#id").val();
-        userEmail = $("#email").val();
-        csrfValue = $("input[name=_csrf]").val();
-        params = {id: userId, email: userEmail, _csrf: csrfValue};
-
-        $.post(url, params, function (response) {
-            if (response === "OK") {
-                form.submit();
-            } else if (response === "Duplicated") {
-                showModal("Warning", "There is another user with the same email " + userEmail);
-            } else {
-                showModal("Error", "Unknown response from server");
-            }
-        }).fail(function () {
-            showModal("Error", "Could not connect to the server");
-        });
-        return false;
-    }
-
-    function showModal(title, message) {
-        $("#modal_title").text(title);
-        $("#modal_body").text(message);
-        $("#modal_dialog").modal('show')
-    }
-</script>
-</html>
