--- ShopmeWebParent/ShopmeBackEnd/src/main/resources/templates/users.html
@@ -1,179 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org"
-      lang="en">
-<head th:replace="~{fragments :: page_head('Users - Shopeme Admin')}"/>
-<body>
-<div class="container-fluid">
-    <div th:replace="~{navigation :: menu}"></div>
-    <div>
-        <h2>Manage Users</h2>
-        <a th:href="@{/users/new}" class="fas fa-user-plus fa-2x icon-dark mr-2"></a>
-        &nbsp;|&nbsp;
-        <a th:href="@{/users/export/csv}" class="fas fa-file-csv fa-2x icon-dark mr-2"></a>
-        &nbsp;|&nbsp;
-        <a th:href="@{/users/export/excel}" class="fas fa-file-excel fa-2x icon-dark"></a>
-        &nbsp;|&nbsp;
-        <a th:href="@{/users/export/pdf}" class="fas fa-file-pdf fa-2x icon-dark"></a>
-    </div>
-    <div th:if="${message != null}" class="alert alert-success text-center">
-        [[${message}]]
-    </div>
-    <div>
-        <form th:action="@{/users/page/1}" class="d-flex flex-row align-items-center flex-wrap m-3">
-            <input type="hidden" name="sortField" th:value="${sortField}"/>
-            <input type="hidden" name="sortDir" th:value="${sortDir}"/>
-            <label class="my-1 me-2" for="keyword">Filter:&nbsp;</label>
-            <input type="search" name="keyword" id="keyword" th:value="${keyword}"
-                   class="form-control my-1 me-sm-2 w-auto" required/>
-            &nbsp;&nbsp;
-            <button type="submit" class="btn btn-primary">
-                <i class="fas fa-search"></i>
-            </button>
-            &nbsp;&nbsp;
-            <button type="button" class="btn btn-secondary" onclick="clearFilter()">
-                <i class="fas fa-eraser"></i>
-            </button>
-        </form>
-    </div>
-
-    <div class="full-details">
-        <table class="table table-bordered table-striped table-hover table-responsive-xl">
-            <thead class="table-dark">
-            <tr>
-                <th class="hideable-column">
-                <th th:replace="~{fragments :: column_link('id', 'User ID', 'tag')}"/>
-                </th>
-
-                <th>Photos</th>
-
-                <th class="hideable-column">
-                <th th:replace="~{fragments :: column_link('email', 'E-mail', 'tag')}"/>
-                </th>
-
-                <th th:replace="~{fragments :: column_link('firstName', 'First Name', 'none')}"/>
-                <th th:replace="~{fragments :: column_link('lastName', 'Last Name', 'none')}"/>
-
-                <th>Roles</th>
-                <th th:replace="~{fragments :: column_link('enabled', 'Enabled', 'none')}"/>
-                <th></th>
-            </tr>
-            </thead>
-            <tbody>
-            <tr th:each="user : ${listUsers}">
-                <td class="hideable-column">[[${user.id}]]</td>
-                <td>
-                    <div th:replace="~{fragments :: user_photos}"/>
-                </td>
-                <td class="hideable-column">[[${user.email}]]</td>
-                <td>[[${user.firstName}]]</td>
-                <td>[[${user.lastName}]]</td>
-                <td>[[${user.roles}]]</td>
-
-                <td>
-                    <div th:replace="~{fragments :: user_status}"/>
-                </td>
-
-                <td>
-                    <div th:replace="~{fragments :: user_actions}"/>
-                </td>
-            </tr>
-            </tbody>
-        </table>
-    </div>
-
-    <div class="less-details">
-        <div th:each="user : ${listUsers}" class="row m-1">
-            <div class="col-4">
-                <div th:replace="~{fragments :: user_photos}"/>
-            </div>
-            <div class="col-8">
-                <div>[[${user.fullName}]]</div>
-                <div>[[${user.roles}]]</div>
-                <div class="mt-2">
-                    <div th:replace="~{fragments :: user_status}"/>
-                    &nbsp;
-                    <div th:replace="~{fragments :: user_actions}"/>
-                </div>
-            </div>
-            <div>&nbsp;</div>
-        </div>
-    </div>
-
-    <div class="text-center m-1" th:if="${totalItems > 0}">
-        <span>Showing user # [[${startCount}]] to [[${endCount}]] of [[${totalItems}]]</span>
-    </div>
-    <div class="text-center m-1" th:if="${totalItems == 0}">
-        <span>No users found</span>
-    </div>
-
-    <div th:if="${totalPages > 1}">
-        <nav>
-            <ul class="pagination justify-content-center flex-wrap">
-                <li th:class="${currentPage > 1 ? 'page-item' : 'page-item disabled'}">
-                    <a th:replace="~{fragments :: page_link(1, 'First')}"></a>
-                </li>
-                <li th:class="${currentPage > 1 ? 'page-item' : 'page-item disabled'}">
-                    <a th:replace="~{fragments :: page_link(${currentPage - 1}, 'Previous')}"></a>
-                </li>
-
-                <li th:class="${currentPage != i ? 'page-item': 'page-item active'}"
-                    th:each="i : ${#numbers.sequence(1, totalPages)}">
-                    <a th:replace="~{fragments :: page_link(${i}, ${i})}"></a>
-                </li>
-
-                <li th:class="${currentPage < totalPages ? 'page-item' : 'page-item disabled'}">
-                    <a th:replace="~{fragments :: page_link(${currentPage + 1}, 'Next')}"></a>
-                </li>
-
-                <li th:class="${currentPage < totalPages ? 'page-item' : 'page-item disabled'}">
-                    <a th:replace="~{fragments :: page_link(${totalPages}, 'Last')}"></a>
-                </li>
-            </ul>
-        </nav>
-    </div>
-    <div class="modal fade text-center" id="delete_dialog">
-        <div class="modal-dialog">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h4 class="modal-title" id="modal_title">Delete Confirmation</h4>
-                </div>
-                <div class="modal-body">
-                    <span id="modal_body"></span>
-                </div>
-                <div class="modal-footer">
-                    <a class="btn btn-success" id="yes_btn" href="">Yes</a>
-                    <button type="button" class="btn btn-danger" data-bs-dismiss="modal">No</button>
-                </div>
-            </div>
-        </div>
-    </div>
-    <div th:replace="~{fragments :: footer}">
-    </div>
-    <script type="text/javascript">
-        $(document).ready(function () {
-            $(".link-delete").on("click", function (e) {
-                e.preventDefault();
-                //alert($(this).attr("href"))
-                const link = $(this);
-                const userId = link.attr("userId");
-                $("#yes_btn").attr("href", link.attr("href"));
-                $("#modal_body").text("Are you sure you want to delete user with ID: " + userId + "?");
-                $("#delete_dialog").modal('show');
-            });
-        });
-
-        function clearFilter() {
-            window.location = "[[@{/users}]]";
-        }
-
-        $(document).ready(function () {
-            $("#logoutLink").on("click", function (e) {
-                e.preventDefault();
-                document.logoutForm.submit();
-            });
-        });
-
-    </script>
-</div>
-</body>
-</html>
