+++ ShopmeWebParent/ShopmeBackEnd/src/main/resources/templates/users/account_form.html
@@ -42,7 +42,7 @@
                     <input type="password" th:field="*{password}" class="form-control"
                            placeholder="Leave blank if you don't want to change password"
                            minlength="8" maxlength="20"
-                            oninput="checkPasswordMatch(document.getElementById('confirmPassword'))"
+                           oninput="checkPasswordMatch(document.getElementById('confirmPassword'))"
                     />
                 </div>
             </div>
@@ -75,64 +75,18 @@
             </div>
         </div>
     </form>
-    <div class="modal fade text-center" id="modal_dialog">
-        <div class="modal-dialog">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h4 class="modal-title" id="modal_title">Warning</h4>
-                    <button type="button" class="btn-close" data-bs-dismiss="modal">&times;</button>
-                </div>
-                <div class="modal-body">
-                    <span id="modal_body"></span>
-                </div>
-
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-danger" data-bs-dismiss="modal">Close</button>
-                </div>
-            </div>
-        </div>
-    </div>
-    <div>
-
-    </div>
     <div th:replace="~{fragments :: footer}"></div>
 </div>
 </body>
 <script type="text/javascript">
-    $(document).ready(function () {
-        $("#btn_cancel").on("click", function () {
-            window.location = "[[@{/users}]]";
-        });
-
-        $("#file_image").change(function () {
-            file_size = this.files[0].size
-            alert("File size: " + file_size);
-            if (file_size > 1048576) {
-                this.setCustomValidity("You must choose an image less than 1MB!");
-                this.reportValidity()
-            } else {
-                this.setCustomValidity("");
-                showImageThumbnail(this);
-            }
-        })
-    });
-
-    function showImageThumbnail(fileInput) {
-        var file = fileInput.files[0];
-        var reader = new FileReader();
-        reader.onload = function (e) {
-            $("#thumbnail").attr("src", e.target.result)
-        };
-
-        reader.readAsDataURL(file);
-    }
-
+    moduleURL = "[[@{/}]]";
     function checkPasswordMatch(confirmPassword) {
-        if (confirmPassword.value != $("#password").val() ) {
+        if (confirmPassword.value != $("#password").val()) {
             confirmPassword.setCustomValidity("Passwords do not match!")
         } else {
             confirmPassword.setCustomValidity("");
         }
     }
 </script>
+<script th:src="@{/js/common_form.js}"></script>
 </html>
